// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** Generic Object Scalar Type */
  BlockAttributesObject: { input: any; output: any; }
};

/** A Field Group registered by ACF */
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type Page = ContentNode & DatabaseIdentifier & HierarchicalContentNode & HierarchicalNode & MenuItemLinkable & Node & NodeWithAuthor & NodeWithComments & NodeWithContentEditor & NodeWithEditorBlocks & NodeWithFeaturedImage & NodeWithPageAttributes & NodeWithPageEditorBlocks & NodeWithRevisions & NodeWithTemplate & NodeWithTitle & Previewable & UniformResourceIdentifiable & {
// ...
  pageBuilderFrontPage?: Maybe<Page_Pagebuilderfrontpage>;
// ...
};
// ...
export type Page_Pagebuilderfrontpage = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage';
  dynamicBlocks?: Maybe<Array<Maybe<Page_Pagebuilderfrontpage_DynamicBlocks>>>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};

export type Page_Pagebuilderfrontpage_DynamicBlocks = Page_Pagebuilderfrontpage_DynamicBlocks_FeaturedItems | Page_Pagebuilderfrontpage_DynamicBlocks_Hero | Page_Pagebuilderfrontpage_DynamicBlocks_Partners | Page_Pagebuilderfrontpage_DynamicBlocks_Subscribenewsletter | Page_Pagebuilderfrontpage_DynamicBlocks_Testimonial | Page_Pagebuilderfrontpage_DynamicBlocks_WhyChooseUs;

/** Group within the flex field */
export type Page_Pagebuilderfrontpage_DynamicBlocks_FeaturedItems = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_FeaturedItems';
  description?: Maybe<Scalars['String']['output']>;
  featuredItemCard?: Maybe<Array<Maybe<Page_Pagebuilderfrontpage_DynamicBlocks_FeaturedItems_FeaturedItemCard>>>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  heading?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Pagebuilderfrontpage_DynamicBlocks_FeaturedItems_FeaturedItemCard = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_FeaturedItems_featuredItemCard';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  heading?: Maybe<Scalars['String']['output']>;
  image?: Maybe<MediaItem>;
};

/** Group within the flex field */
export type Page_Pagebuilderfrontpage_DynamicBlocks_Hero = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_Hero';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  image?: Maybe<MediaItem>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Group within the flex field */
export type Page_Pagebuilderfrontpage_DynamicBlocks_Partners = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_Partners';
  branchLogos?: Maybe<Array<Maybe<Page_Pagebuilderfrontpage_DynamicBlocks_Partners_BranchLogos>>>;
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  heading?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Pagebuilderfrontpage_DynamicBlocks_Partners_BranchLogos = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_Partners_branchLogos';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  logo?: Maybe<MediaItem>;
};

/** Group within the flex field */
export type Page_Pagebuilderfrontpage_DynamicBlocks_Subscribenewsletter = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_Subscribenewsletter';
  background?: Maybe<MediaItem>;
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  heading?: Maybe<Scalars['String']['output']>;
};

/** Group within the flex field */
export type Page_Pagebuilderfrontpage_DynamicBlocks_Testimonial = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_Testimonial';
  card?: Maybe<Array<Maybe<Page_Pagebuilderfrontpage_DynamicBlocks_Testimonial_Card>>>;
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  heading?: Maybe<Scalars['String']['output']>;
  sectionTitle?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Pagebuilderfrontpage_DynamicBlocks_Testimonial_Card = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_Testimonial_card';
  avatar?: Maybe<MediaItem>;
  comment?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  ratingScore?: Maybe<Scalars['Float']['output']>;
  reviewer?: Maybe<Scalars['String']['output']>;
  reviewerTitle?: Maybe<Scalars['String']['output']>;
};

/** Group within the flex field */
export type Page_Pagebuilderfrontpage_DynamicBlocks_WhyChooseUs = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_WhyChooseUs';
  background?: Maybe<MediaItem>;
  card?: Maybe<Array<Maybe<Page_Pagebuilderfrontpage_DynamicBlocks_WhyChooseUs_Card>>>;
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  heading?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Pagebuilderfrontpage_DynamicBlocks_WhyChooseUs_Card = AcfFieldGroup & {
  __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_WhyChooseUs_card';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  heading?: Maybe<Scalars['String']['output']>;
  icon?: Maybe<MediaItem>;
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type RootQuery = {
// ...
  page?: Maybe<Page>;
// ...
};
// ...
export type GetHomePageQuery = { __typename?: 'RootQuery', page?: { __typename?: 'Page', title?: string | null, pageBuilderFrontPage?: { __typename?: 'Page_Pagebuilderfrontpage', dynamicBlocks?: Array<{ __typename: 'Page_Pagebuilderfrontpage_DynamicBlocks_FeaturedItems', heading?: string | null, description?: string | null, featuredItemCard?: Array<{ __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_FeaturedItems_featuredItemCard', heading?: string | null, description?: string | null, image?: { __typename?: 'MediaItem', sizes?: string | null, altText?: string | null, sourceUrl?: string | null } | null } | null> | null } | { __typename: 'Page_Pagebuilderfrontpage_DynamicBlocks_Hero', title?: string | null, description?: string | null, image?: { __typename?: 'MediaItem', altText?: string | null, sourceUrl?: string | null } | null } | { __typename: 'Page_Pagebuilderfrontpage_DynamicBlocks_Partners', heading?: string | null, description?: string | null, branchLogos?: Array<{ __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_Partners_branchLogos', logo?: { __typename?: 'MediaItem', altText?: string | null, sourceUrl?: string | null } | null } | null> | null } | { __typename: 'Page_Pagebuilderfrontpage_DynamicBlocks_Subscribenewsletter', heading?: string | null, description?: string | null, background?: { __typename?: 'MediaItem', altText?: string | null, sourceUrl?: string | null } | null } | { __typename: 'Page_Pagebuilderfrontpage_DynamicBlocks_Testimonial', sectionTitle?: string | null, heading?: string | null, description?: string | null, card?: Array<{ __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_Testimonial_card', reviewer?: string | null, reviewerTitle?: string | null, ratingScore?: number | null, comment?: string | null, avatar?: { __typename?: 'MediaItem', altText?: string | null, sourceUrl?: string | null } | null } | null> | null } | { __typename: 'Page_Pagebuilderfrontpage_DynamicBlocks_WhyChooseUs', heading?: string | null, description?: string | null, background?: { __typename?: 'MediaItem', altText?: string | null, sourceUrl?: string | null } | null, card?: Array<{ __typename?: 'Page_Pagebuilderfrontpage_DynamicBlocks_WhyChooseUs_card', heading?: string | null, description?: string | null, url?: string | null, icon?: { __typename?: 'MediaItem', altText?: string | null, sourceUrl?: string | null } | null } | null> | null } | null> | null } | null } | null };